temp(trigger):完成部分文章接口

+++ blog-infrastructure/src/main/java/com/wbxnl/blog/infrastructure/persistent/utils/PageUtils.java
@@ -28,7 +28,7 @@ public class PageUtils {
         PageData<T> pageData = new PageData<>();
         pageData.setNumber(number);
         pageData.setSize(size);
-        pageData.setTotal((int) total);
+        pageData.setTotal(total);
         pageData.setData(data);
         return pageData;
     }
